Rewrote game loop

+++ ClosedGL/include/ClosedGL/Scene/PerspectiveCamera.cpp
@@ -25,9 +25,10 @@ void PerspectiveCamera::update(Timestep ts) {
 			mView = glm::translate(mView, glm::vec3(-mSpeed * ts, 0, 0));
 
 		if (Input::isKeyDown(Keys::Q))
-			mView = glm::rotate(mView, mSpeed * ts, glm::vec3(0, mSpeed * ts, 0));
+			mView = glm::rotate(mView, mSpeed * (float) ts, glm::vec3(0, mSpeed * ts, 0));
 		if (Input::isKeyDown(Keys::E))
-			mView = glm::rotate(mView, mSpeed * ts, glm::vec3(0, -mSpeed * ts, 0));
+			mView = glm::rotate(mView, mSpeed * (float) ts, glm::vec3(0, -mSpeed * ts, 0));
+	
 	}
 
 }
